- Check for disk space and if there is insufficient disk space on server then it will display proper message.
git-svn-id: http://svn.ez.no/svn/extensions/oo/trunk@137 49f50cf4-cce6-0310-bc1e-a21d999f3d8f

+++ deamon.php
@@ -66,6 +66,11 @@ function convert_to( $fileName, $convertCommand, $tmpFile )
 
     $tmpFile = $tmpPath . $tmpFile;
 
+    if( filesize( $tmpFile ) >= disk_free_space("/") )
+    {
+        socket_write( $spawn, "Error: (3)-Not Enough Disk space." );
+        return false;
+    }
     unlink( $tmpFile );
     $result = shell_exec( $ooexecutable . " -writer 'macro:///standard.Module1." . $convertCommand . "($fileName)'" );
